backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
requests
numpy
pandas
scikit-learn
python-dotenv
openai
gunicorn
